// backend/src/models/InstagramModel.ts
export interface InstagramData {
    data: Data;
}

interface Data {
    accessibility_caption: string | null;
    are_remixes_crosspostable: boolean;
    avatar_stickers: any[];
    boost_unavailable_identifier: string | null;
    boost_unavailable_reason: string | null;
    boost_unavailable_reason_v2: string | null;
    can_reply: boolean;
    can_reshare: boolean;
    can_save: boolean;
    caption: Caption;
    caption_is_edited: boolean;
    clips_metadata: any;
    clips_tab_pinned_user_ids: any[];
    coauthor_producer_can_see_organic_insights: boolean;
    coauthor_producers: any[];
    code: string;
    comment_inform_treatment: any;
    comment_threading_enabled: boolean;
    comments_disabled: boolean | null;
    commerce_integrity_review_decision: string;
    commerciality_status: string;
    creative_config: any;
    creator_viewer_insights: any[];
    crosspost_metadata: any;
    deleted_reason: number;
    device_timestamp: number;
    fb_aggregated_comment_count: number;
    fb_aggregated_like_count: number;
    fbid: string;
    featured_products: any[];
    filter_type: number;
    fundraiser_tag: any;
    gen_ai_detection_method: any;
    has_audio: boolean;
    has_high_risk_gen_ai_inform_treatment: boolean;
    has_liked: boolean;
    has_more_comments: boolean;
    has_privately_liked: boolean;
    has_shared_to_fb: number;
    has_views_fetching: boolean;
    id: string;
    ig_media_sharing_disabled: boolean;
    ig_play_count: number;
    igbio_product: any | null;
    image_versions: any;
    inline_composer_display_condition: string;
    inline_composer_imp_trigger_time: number;
    integrity_review_decision: string;
    invited_coauthor_producers: any[];
    is_artist_pick: boolean;
    is_comments_gif_composer_enabled: boolean;
    is_cutout_sticker_allowed: boolean;
    is_dash_eligible: number;
    is_eligible_for_media_note_recs_nux: boolean;
    is_eligible_for_meta_ai_share: boolean;
    is_in_profile_grid: boolean;
    is_open_to_public_submission: boolean;
    is_organic_product_tagging_eligible: boolean;
    is_paid_partnership: boolean;
    is_pinned: boolean;
    is_post_live_clips_media: boolean;
    is_quiet_post: boolean;
    is_reshare_of_text_post_app_media_in_ig: boolean;
    is_reuse_allowed: boolean;
    is_social_ufi_disabled: boolean;
    is_tagged_media_shared_to_viewer_profile_grid: boolean;
    is_third_party_downloads_eligible: boolean;
    is_unified_video: boolean;
    is_video: boolean;
    like_and_view_counts_disabled: boolean;
    location: any | null;
    max_num_visible_preview_comments: number;
    media_cropping_info: any;
    media_name: string;
    media_notes: any;
    media_type: number;
    meta_ai_suggested_prompts: any[];
    metrics: any;
    music_metadata: any | null;
    number_of_qualities: number;
    open_carousel_show_follow_button: boolean;
    original_height: number;
    original_width: number;
    owner: any;
    pk: string;
    preview_comments: any[];
    product_suggestions: any[];
    product_type: string;
    share_count_disabled: boolean;
    sharing_friction_info: any;
    shop_routing_user_id: any | null;
    should_show_author_pog_for_tagged_media_shared_to_profile_grid: boolean;
    subscribe_cta_visible: boolean;
    tagged_users: any | null;
    taken_at: number;
    thumbnail_url: string;
    timeline_pinned_user_ids: any[];
    top_likers: any[];
    user: any;
    video_codec: string;
    video_duration: number;
    video_sticker_locales: any[];
    video_subtitles_confidence: number;
    video_subtitles_locale: string;
    video_url: string;
    video_versions: any[];
}

interface Caption {
    created_at: number;
    created_at_utc: number;
    did_report_as_spam: boolean;
    hashtags: any[];
    id: string;
    is_covered: boolean;
    is_ranked_comment: boolean;
    mentions: any[];
    pk: string;
    private_reply_status: number;
    share_enabled: boolean;
    text: string;
    type: number;
}